Main class changes, command moving part 2
