+++ src/main/java/me/partlysanestudios/partlysaneskies/features/commands/Discord.java
@@ -1,2 +1,29 @@
-package me.partlysanestudios.partlysaneskies.features.commands;public class Discord {
+package me.partlysanestudios.partlysaneskies.features.commands;
+
+import me.partlysanestudios.partlysaneskies.PartlySaneSkies;
+import me.partlysanestudios.partlysaneskies.commands.PSSCommand;
+import net.minecraft.event.ClickEvent;
+import net.minecraft.util.ChatComponentText;
+import net.minecraft.util.IChatComponent;
+
+public class Discord {
+    public static void registerDiscordCommand() {
+        new PSSCommand("discord")
+                .addAlias("pssdisc")
+                .addAlias("pssd")
+                .addAlias("psdisc")
+                .addAlias("psdiscord")
+                .addAlias("pssdiscord")
+                .addAlias("didcord")
+                .setDescription("Join the Partly Sane Studios Discord Server")
+                .setRunnable((s, a) -> {
+                    // Creates a new message with the correct text
+                    IChatComponent message = new ChatComponentText(PartlySaneSkies.CHAT_PREFIX + "§9Join the discord: https://discord.gg/" + PartlySaneSkies.discordCode);
+                    // Sets the text to be clickable with a link
+                    message.getChatStyle().setChatClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, "https://discord.gg/" + PartlySaneSkies.discordCode));
+                    // Prints message
+                    PartlySaneSkies.minecraft.ingameGUI.getChatGUI().printChatMessage(message);
+                })
+                .register();
+    }
 }
