+++ src/main/java/me/partlysanestudios/partlysaneskies/features/help/HelpCommands.java
@@ -23,26 +23,6 @@ import java.util.Map;
 
 public class HelpCommands {
 
-    public static void registerDiscordCommand() {
-        new PSSCommand("discord")
-                .addAlias("pssdisc")
-                .addAlias("pssd")
-                .addAlias("psdisc")
-                .addAlias("psdiscord")
-                .addAlias("pssdiscord")
-                .addAlias("didcord")
-                .setDescription("Join the Partly Sane Studios Discord Server")
-                .setRunnable((s, a) -> {
-                    // Creates a new message with the correct text
-                    IChatComponent message = new ChatComponentText(PartlySaneSkies.CHAT_PREFIX + "§9Join the discord: https://discord.gg/" + PartlySaneSkies.discordCode);
-                    // Sets the text to be clickable with a link
-                    message.getChatStyle().setChatClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, "https://discord.gg/" + PartlySaneSkies.discordCode));
-                    // Prints message
-                    PartlySaneSkies.minecraft.ingameGUI.getChatGUI().printChatMessage(message);
-                })
-                .register();
-    }
-
     public static void registerConfigCommand() {
         new PSSCommand("pssconfig")
                 .addAlias("pssc")
@@ -68,58 +48,6 @@ public class HelpCommands {
                 }).register();
     }
 
-    public static void registerVersionCommand(){
-        new PSSCommand("partlysaneskiesversion")
-                .addAlias("pssversion")
-                .addAlias("pssv")
-                .addAlias("partlysaneskiesv")
-                .setDescription("Prints the version of Partly Sane Skies you are using")
-                .setRunnable((s, a) -> {
-                    ChatComponentText chatcomponent = new ChatComponentText("§b§m-----------------------------------------------------§0" +
-                        "\n§b§lPartly Sane Skies Version:" +
-                        "\n    §e" + PartlySaneSkies.VERSION +	
-                        (PartlySaneSkies.isLatestVersion() ? "\n§aYou are using the latest version of Partly Sane Skies!" : "\n§cYou are not using the latest version of Partly Sane Skies! Click here to download the newest version!") +
-                        "\n§b§m-----------------------------------------------------§0"
-                    );
-                    if (!PartlySaneSkies.isLatestVersion()){
-                        chatcomponent.getChatStyle().setChatClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, "https://github.com/PartlySaneStudios/partly-sane-skies/releases"));
-                    }
-
-                    ChatUtils.INSTANCE.sendClientMessage(chatcomponent);
-                }).register();
-    }
-
-    public static void registerCrepesCommand() {
-        new PSSCommand("crepes")
-                .addAlias("crêpes")
-                .setDescription("Crepes!")
-                .setRunnable((s, a) -> {
-                    ChatComponentText chatComponent = new ChatComponentText("§0§m-----------------------------------------------------§0" +
-                            "\n§6Ingredients:" +
-                            "\n" +
-                            "\n§f- 250g of wheat flour" +
-                            "\n§f- 50g of butter" +
-                            "\n§f- 50cl of milk" +
-                            "\n§f- 10cl of water" +
-                            "\n§f- 4 (big) eggs" +
-                            "\n§f- 2 tablespoons of powdered sugar" +
-                            "\n§f- 1 pinch of salt" +
-                            "\n" +
-                            "\n§f1) In a large mixing bowl, add the wheat flour, the 2 tablespoons of powdered sugar, and 1 pinch of salt" +
-                            "\n§f2) While whisking the bowl, progressively pour in the milk and the water " +
-                            "\n§f3) After, add in the (melted) butter and the eggs" +
-                            "\n§f4) Leave the bowl in the fridge for 30-45 minutes" +
-                            "\n§f5) Look up a tutorial on how to make crêpes on a pan" +
-                            "\n§f6) Enjoy your (burnt) crêpes with something like Nutella, sugar, or marmalade" +
-                            "\n" +
-                            "\n§fSource: " +
-                            "\n§5http://www.recettes-bretonnes.fr/crepe-bretonne/recette-crepe.html" +
-                            "\n§0§m-----------------------------------------------------§0");
-                    chatComponent.getChatStyle().setChatClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL,"http://www.recettes-bretonnes.fr/crepe-bretonne/recette-crepe.html"));
-                    ChatUtils.INSTANCE.sendClientMessage(chatComponent);
-                }).register();
-
-    }
     static List<String> configAliases = Arrays.asList("conf", "c", "config");
     public static void registerHelpCommand() {
         new PSSCommand("psshelp")
